requirements.txt
asgiref==3.10.0
blinker==1.9.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.2.1
colorama==0.4.6
Django==5.2.7
django-cors-headers==4.9.0
djangorestframework==3.16.1
dnspython==2.7.0
Faker==37.4.0
Flask==3.1.2
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pillow==12.0.0
pymongo==4.14.1
requests==2.32.5
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
